Move things around so specs pass again.
